requirements.txt
groq
pandas
pytest
streamlit
python-dotenv
boto3
google-cloud-secret-manager
detect-secrets
